refactor: historyAPI 위치 변경
